1. Adding beam search for validation during training 2. Output attention weights from multi-heads attention layer

+++ Seq2SeqSharp/Utils/Vocab.cs
@@ -229,7 +229,7 @@ namespace Seq2SeqSharp
         }
 
 
-        public List<string> ConvertTargetIdsToString(List<int> idxs)
+        public List<string> ConvertTargetIdsToString(List<float> idxs)
         {
             lock (locker)
             {
